# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SilentRelay API
  description: |
    Secure end-to-end encrypted messaging API for SilentRelay.
    
    ## Authentication
    Most endpoints require JWT authentication via `Authorization: Bearer <token>` header.
    
    ## Rate Limiting
    - SMS endpoints: 5 requests/minute/IP
    - Auth endpoints: 10 requests/minute/IP  
    - Protected endpoints: 100 requests/minute/user
    
    ## Security
    All communications use end-to-end encryption. The server never has access to message content.
  version: 1.0.0
  contact:
    name: SilentRelay Support
    url: https://silentrelay.com.au
  license:
    name: Proprietary
    url: https://silentrelay.com.au/terms
servers:
  - url: https://silentrelay.com.au/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or registration
  schemas:
...
    PrivacySettings:
      type: object
      properties:
        profile_photo:
          type: string
          enum: [everyone, contacts, nobody]
        last_seen:
          type: string
          enum: [everyone, contacts, nobody]
        read_receipts:
          type: boolean
paths:
...
  /media/upload-url:
    post:
      tags:
        - Media
      summary: Get upload URL
      description: Generates a presigned URL for media upload
      operationId: getUploadUrl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_name
                - content_type
                - size
              properties:
                file_name:
                  type: string
                content_type:
                  type: string
                  example: "image/jpeg"
                size:
                  type: integer
                  description: File size in bytes
                  maximum: 104857600
      responses:
        '200':
          description: Upload URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_url:
                    type: string
                    format: uri
                  media_id:
                    type: string
                    format: uuid
                  expires_at:
                    type: string
                    format: date-time
  /media/{mediaId}:
    get:
      tags:
        - Media
      summary: Get media URL
      description: Generates a presigned URL for media download
      operationId: getMediaUrl
      security:
        - bearerAuth: []
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  download_url:
                    type: string
                    format: uri
                  expires_at:
                    type: string
                    format: date-time
        '404':
          description: Media not found

  # =====================
  # PRIVACY
  # =====================
  /privacy:
    get:
      tags:
        - Privacy
      summary: Get privacy settings
      operationId: getPrivacySettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Privacy settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacySettings'

    post:
      tags:
        - Privacy
      summary: Update privacy setting
      operationId: updatePrivacySetting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                setting:
                  type: string
                  enum: [profile_photo, last_seen, read_receipts]
                value:
                  type: string
      responses:
        '200':
          description: Setting updated
...